+++ package.json
@@ -13,8 +13,7 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "api": "node src/API/api.js",
-    "launch": "react-scripts start && node src/api.js"
+    "api": "node src/API/server.js",
   },
   "eslintConfig": {
     "extends": "react-app"
