+++ src/API/api.js
@@ -1,8 +1,24 @@
 const express = require('express')
-const app = express()
-const port = 4000
-const api = require('./router.js')
+const router = express.Router()
+const displayLogTime = require('./Utils/timeHelpers.js')
+// DATA
+const NortonRooms = [011, 012, 013, 015, 016, 017, 020, 101, 102, 103, 104, 105, 201, 202, 203, 204, 205, 206, 207, 208, 209, 232]
+const CookeRooms = [8, 221, 224, 'CCRH', 'Heeren']
+const CarverRooms = [108, 135, 'Ingram']
+const LibraryRooms = ['Mullins', 'Crismon', 'Curriculum']
 
-app.use("/api", api)
+router.use((req, res, next) => {
+	let logTime = new Date()
+	console.log(displayLogTime())
+	next()
+})
+router.get("/buildings/:building", (req, res) => {
+	const building = req.params.building
+	res.send(`You requested building:${building}`)
+})
+router.get("/rooms/:room", (req, res) => {
+	const room = req.params.room
+	res.send(`You requested room ${room}`) 
+})
 
-app.listen(port, () => console.log(`Listening on port ${port}`))
+module.exports = router
